Move DatasetExport into its own dataset_export.py module
Relocate the generic DatasetExport base out of image_dataset_export.py into a
new dataset_export.py. image_dataset_export.py keeps the public ImageDatasetExport
and image_sample_to_image (so lightly_studio.export.image_dataset_export still
imports) and now imports DatasetExport from the new module.

Pure move, no behavior change. The only follow-on edit is retargeting the
COCOObjectDetectionOutput writer patch in the test to the module it now lives in.
